Delete README.md
I have used LDR-TORCH, ARDUINO-UNO, LED LIGHT, LCD SCREEN, RESISTORS

This works similarly to the street solar light, where solar light glows when light intensity decreases.
It works similar to that but the difference is that it works precisely and used in the house,office etc.
In this circuit as the intensity of  natural light decreases the led bulb start lighting up but not all at the same time, slowly slowly the number of 
led light which glows increases as the natural light intensity decreases. After a certain limit when natural light becomes very dim or zero , all the led light up.